Fixed initialization for adc124
The strain gauge readers should now work!
The problem was an unknown frame offset due to not having control of the
Chip Select pin. So, the frame offset is calibrated off channel 4, which
should always have a pull-up resistor and never be connected to a sensor.

If this has to change, there are other ways we can do this, but they require
changes in the hardware interface.
- Marshal